Mark old triggers as legacy
