+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "linear-zapier",
-  "version": "4.8.5",
+  "version": "4.8.6",
   "description": "Linear's Zapier integration",
   "main": "index.js",
   "license": "MIT",
