+++ src/triggers/projectUpdate.ts
@@ -1,176 +1,170 @@
-import sample from "../samples/projectUpdate.json";
+import { pick } from "lodash";
 import { ZObject, Bundle } from "zapier-platform-core";
+import sample from "../samples/projectUpdate.json";
+import { getWebhookData, unsubscribeHook } from "../handleWebhook";
+import { jsonToGraphQLQuery, VariableType } from "json-to-graphql-query";
+import { fetchFromLinear } from "../fetchFromLinear";
+
+interface ProjectUpdate {
+  id: string;
+  body: string;
+  url: string;
+  createdAt: Date;
+  updatedAt: Date;
+  editedAt: Date | null;
+  health: string;
+  project: {
+    id: string;
+    name: string;
+  };
+  user: {
+    id: string;
+    name: string;
+    email: string;
+  };
+}
 
 interface ProjectUpdatesResponse {
   data: {
     projectUpdates: {
-      nodes: {
-        id: string;
-        body: string;
-        diffMarkdown: string;
-        createdAt: Date;
-        updatedAt: Date;
-        project?: {
-          id: string;
-          name: string;
-        };
-        user: {
-          id: string;
-          name: string;
-          email: string;
-        };
-      }[];
-      pageInfo: {
-        hasNextPage: boolean;
-        endCursor: string;
-      };
+      nodes: ProjectUpdate[];
     };
   };
 }
 
-const buildProjectUpdateList = (orderBy: "createdAt" | "updatedAt") => async (z: ZObject, bundle: Bundle) => {
-  const cursor = bundle.meta.page ? await z.cursor.get() : undefined;
+type EventType = "create" | "update";
 
-  const response = await z.request({
-    url: "https://api.linear.app/graphql",
-    headers: {
-      "Content-Type": "application/json",
-      Accept: "application/json",
-      authorization: bundle.authData.api_key,
-    },
-    body: {
-      query: `
-      query ZapierListProjectUpdates(
-        $after: String
-        $teamId: ID
-        $projectId: ID
-        $userId: ID
-        $orderBy: PaginationOrderBy!
-      ) {
-        projectUpdates(
-          first: 10
-          after: $after
-          orderBy: $orderBy
-          filter: {
-            project: {
-              id: { eq: $projectId }
-              accessibleTeams: { id: { eq: $teamId } }
-            }
-            user: { id: { eq: $userId } }
-          }
-        ) {
-          nodes {
-            id
-            body
-            diffMarkdown
-            health
-            url
-            editedAt
-            createdAt
-            updatedAt
-            project {
-              id
-              name
-            }
-            user {
-              id
-              name
-              email
-            }
-          }
-          pageInfo {
-            hasNextPage
-            endCursor
-          }
-        }
-      }
-      `,
-      variables: {
-        after: cursor,
-        teamId: bundle.inputData.team_id,
-        projectId: bundle.inputData.project_id,
-        userId: bundle.inputData.user_id,
-        orderBy,
-      },
-    },
-    method: "POST",
-  });
+const subscribeHook = (eventType: EventType) => async (z: ZObject, bundle: Bundle) => {
+  const data = {
+    url: bundle.targetUrl,
+    inputData:
+      bundle.inputData && Object.keys(bundle.inputData).length > 0
+        ? pick(bundle.inputData, ["creatorId", "projectId", "teamId"])
+        : undefined,
+  };
 
-  const data = (response.json as ProjectUpdatesResponse).data;
-  const projectUpdates = data.projectUpdates.nodes;
+  const webhookType = eventType === "create" ? "createProjectUpdate" : "updateProjectUpdate";
+  return z
+    .request({
+      url: `https://client-api.linear.app/connect/zapier/subscribe/${webhookType}`,
+      method: "POST",
+      body: data,
+    })
+    .then((response) => response.data);
+};
 
-  // Set cursor for pagination
-  if (data.projectUpdates.pageInfo.hasNextPage) {
-    await z.cursor.set(data.projectUpdates.pageInfo.endCursor);
+const getProjectUpdateList = () => async (z: ZObject, bundle: Bundle) => {
+  const variables: Record<string, string> = {};
+  const variableSchema: Record<string, string> = {};
+  const filters: unknown[] = [];
+  if (bundle.inputData.creatorId) {
+    variableSchema.creatorId = "ID";
+    variables.creatorId = bundle.inputData.creatorId;
+    filters.push({ user: { id: { eq: new VariableType("creatorId") } } });
   }
+  if (bundle.inputData.teamId) {
+    variableSchema.teamId = "ID";
+    variables.teamId = bundle.inputData.teamId;
+    filters.push({ project: { accessibleTeams: { id: { eq: new VariableType("teamId") } } } });
+  }
+  if (bundle.inputData.projectId) {
+    variableSchema.projectId = "ID";
+    variables.projectId = bundle.inputData.projectId;
+    filters.push({ project: { id: { eq: new VariableType("projectId") } } });
+  }
+  const filter = { and: filters };
 
-  return projectUpdates.map((projectUpdate) => ({
-    ...projectUpdate,
-    id: `${projectUpdate.id}-${projectUpdate[orderBy]}`,
-    projectUpdateId: projectUpdate.id,
-  }));
+  const jsonQuery = {
+    query: {
+      __variables: variableSchema,
+      projectUpdates: {
+        __args: {
+          first: 25,
+          filter,
+        },
+        nodes: {
+          id: true,
+          body: true,
+          url: true,
+          createdAt: true,
+          updatedAt: true,
+          editedAt: true,
+          health: true,
+          project: {
+            id: true,
+            name: true,
+          },
+          user: {
+            id: true,
+            name: true,
+            email: true,
+          },
+        },
+      },
+    },
+  };
+  const query = jsonToGraphQLQuery(jsonQuery);
+  const response = await fetchFromLinear(z, bundle, query, variables);
+  const data = (response.json as ProjectUpdatesResponse).data;
+  return data.projectUpdates.nodes;
 };
 
-const projectUpdate = {
+const operationBase = {
+  inputFields: [
+    {
+      required: false,
+      label: "Creator",
+      key: "creatorId",
+      helpText: "Only trigger on project updates added by this user.",
+      dynamic: "user.id.name",
+      altersDynamicFields: true,
+    },
+    {
+      required: false,
+      label: "Project",
+      key: "projectId",
+      helpText: "Only trigger on project updates tied to this project.",
+      dynamic: "projectWithoutTeam.id.name",
+      altersDynamicFields: true,
+    },
+    {
+      required: false,
+      label: "Team",
+      key: "teamId",
+      helpText: "Only trigger on project updates created in projects tied to this team.",
+      dynamic: "team.id.name",
+      altersDynamicFields: true,
+    },
+  ],
+  type: "hook",
+  perform: getWebhookData,
+  performUnsubscribe: unsubscribeHook,
+  performList: getProjectUpdateList(),
+  sample,
+};
+
+export const newProjectUpdateInstant = {
   noun: "Project Update",
-
-  operation: {
-    inputFields: [
-      {
-        required: false,
-        label: "Team",
-        key: "team_id",
-        helpText: "Limit to project updates from projects in this team.",
-        dynamic: "team.id.name",
-        altersDynamicFields: true,
-      },
-      {
-        required: false,
-        label: "Project",
-        key: "project_id",
-        helpText: "Project this update belongs to.",
-        dynamic: "project.id.name",
-        altersDynamicFields: true,
-      },
-      {
-        required: false,
-        label: "Creator",
-        key: "user_id",
-        helpText: "The user who created this project update.",
-        dynamic: "user.id.name",
-        altersDynamicFields: true,
-      },
-    ],
-    sample,
-  },
-};
-
-export const newProjectUpdate = {
-  ...projectUpdate,
-  key: "newProjectUpdate",
+  key: "newProjectUpdateInstant",
   display: {
     label: "New Project Update",
     description: "Triggers when a new project update is created.",
-    hidden: true,
   },
   operation: {
-    ...projectUpdate.operation,
-    perform: buildProjectUpdateList("createdAt"),
-    canPaginate: true,
+    ...operationBase,
+    performSubscribe: subscribeHook("create"),
   },
 };
 
-export const updatedProjectUpdate = {
-  ...projectUpdate,
-  key: "updatedProjectUpdate",
+export const updatedProjectUpdateInstant = {
+  noun: "Project Update",
+  key: "updatedProjectUpdateInstant",
   display: {
     label: "Updated Project Update",
     description: "Triggers when a project update is updated.",
-    hidden: true,
   },
   operation: {
-    ...projectUpdate.operation,
-    perform: buildProjectUpdateList("updatedAt"),
-    canPaginate: true,
+    ...operationBase,
+    performSubscribe: subscribeHook("update"),
   },
 };
